% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bplot.R
\name{bplot_scmarkers}
\alias{bplot_scmarkers}
\title{bplot_scmarkers}
\usage{
bplot_scmarkers(
  scRNA,
  label = "seurat_clusters",
  width = 20,
  height = 8,
  Neutrophils = NA,
  Basophils = NA,
  Eosinophils = NA,
  Masts = c("TPSAB1", "TPSB2", "CPA3", "FCER1A", "KIT", "MS4A2", "GATA2"),
  DCs = c("CLEC10A", "LAMP3", "ITGAX"),
  cDC1s = c("CLEC9A"),
  cDC2s = c("CD1C"),
  pDCs = c("CLEC4C"),
  Plasmas = c("IGHG1", "MZB1", "TNFRSF17", "SLAMF7", "XBP1"),
  Bs = c("CD79A", "CD79B"),
  Ts = c("CD3D", "CD3E", "CD3G"),
  MAITs = c("TRAV1-2", "KLRB1", "ZBTB16"),
  gdTs = c("TRDC", "TRGC"),
  CD4s = c("CD4"),
  CD4Tfhs = c("CXCR5", "ICOS", "BCL6"),
  CD4Th1s = c("TBX21"),
  CD4Th2s = c("GATA3", "IL4", "IL5", "IL13"),
  CD4Th9s = c("SPI1", "IL9"),
  CD4Th17s = c("RORC", "IL17A", "IL17F", "IL23R"),
  CD4Tr1s = c("IL10"),
  CD8s = c("CD8A", "CD8B"),
  Texs = c("LAG3", "TIGIT", "PDCD1", "HAVCR2"),
  Tregs = c("FOXP3", "CTLA4", "IL2RA"),
  Tns = c("TCF7", "SELL", "LEF1", "CCR7"),
  Teffs = c("KLRG1", "NKG7", "GNLY", "GZMB", "PRF1", "IFNG"),
  Tsts = c("HSPA1A", "HSPA1B", "HSPH1"),
  Tcms = c("FAS", "CXCR3"),
  Tems = NA,
  NKs = c("NCR1", "KLRC1", "KIR2DL4", "XCL1", "XCL2", "IL18R1", "KLRD1", "NCAM1"),
  Monocytes = c("CD14", "FCGR3A", "S100A12", "CD300E", "FCN1"),
  Macrophages = c("LYZ", "CD68", "C1QC", "CSF1R", "AIF1", "ALB"),
  M1s = c("NOS2"),
  M2s = c("MCR1", "ARG1"),
  Cardiomyocytes = c("ACTC1", "MYH7", "TNNT2", "TNNI3", "TTN", "ACTN2", "MYH6"),
  Fibroblasts = c("DCN", "COL1A1", "COL1A2", "THY1"),
  Endothelials = c("TEK", "PECAM1", "FLT1", "VWF"),
  Mesothelials = c("MSLN", "UPK3B", "CALB2", "WT1"),
  SMCs = c("ACTA2", "MYH11"),
  Epithelials = c("EPCAM", "KRT19", "CDH1", "KRT18"),
  Prolifs = c("MKI67"),
  Mal_PNETs = c("SYP", "CHGA"),
  Mals = NA,
  name_use = "Cell",
  mycolor = c("#BC3C29FF", "#0072B5FF", "#E18727FF", "#20854EFF", "#7876B1FF",
    "#6F99ADFF", "#FFDC91FF", "#EE4C97FF", "#E64B35FF", "#4DBBD5FF", "#00A087FF",
    "#3C5488FF", "#F39B7FFF", "#8491B4FF", "#91D1C2FF", "#DC0000FF", "#7E6148FF",
    "#B09C85FF", "#3B4992FF", "#EE0000FF", "#008B45FF", "#631879FF", "#008280FF",
    "#BB0021FF", "#5F559BFF", "#A20056FF", "#808180FF", "#00468BFF", "#ED0000FF",
    "#42B540FF", "#0099B4FF", "#925E9FFF", "#FDAF91FF", "#AD002AFF", "#ADB6B6FF",
    "#374E55FF", "#DF8F44FF", "#00A1D5FF", "#B24745FF", 
     "#79AF97FF", "#6A6599FF",
    "#80796BFF", "#1f77b4", "#ff7f0e", "#279e68", "#d62728", "#aa40fc", "#8c564b",
    "#e377c2", "#b5bd61", "#17becf", "#aec7e8")
)
}
\arguments{
\item{scRNA}{Seurat object}

\item{label}{celltype or seurat_clusters or others}

\item{width}{pdf width}

\item{height}{pdfheight}

\item{mycolor}{mycolor}
}
\value{
pdf and ggplot object
}
\description{
bplot scmarkers
}
