% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bplot.R
\name{bplot_scmarkers_mouse}
\alias{bplot_scmarkers_mouse}
\title{bplot_scmarkers_mouse}
\usage{
bplot_scmarkers_mouse(
  scRNA,
  label = "seurat_clusters",
  width = 20,
  height = 8,
  Neutrophils = c("Ly6g", "S100a8", "S100a9", "Elane"),
  Basophils = c("Prss34", "Mcpt8"),
  Eosinophils = c("Prg2", "Epx"),
  Masts = c("Cpa3", "Mcpt4", "Fcer1a", "Kit"),
  DCs = c("Cd209a", "Zbtb46", "Flt3", "Itgax", "Itgae"),
  cDC1s = c("Batf3", "Clec9a", "Irf8", "Xcr1"),
  cDC2s = c("Sirpa", "Itgam"),
  pDCs = c("Siglech", "Bst2"),
  Plasmas = c("Sdc1", "Jchain", "Tnfrsf17", "Xbp1"),
  Bs = c("Cd79a", "Cd79b", "Cd19", "Ms4a1"),
  Ts = c("Cd3d", "Cd3e", "Cd3g"),
  CD4s = c("Cd4"),
  CD8s = c("Cd8a", "Cd8b1"),
  Texs = c("Lag3", "Tigit", "Havcr2", "Pdcd1"),
  Tregs = c("Foxp3", "Ctla4"),
  Tns = c("Tcf7", "Sell", "Lef1", "Ccr7"),
  Teffs = c("Ifng", "Tnf", "Prf1", "Gzma", "Gzmb", "Il2ra", "Il2rb"),
  NKs = c("Ncr1", "Ncam1", "Klrb1c", "Klrd1", "Klra8", "Klrk1"),
  Monocytes = c("Ly6c2", "Ly6c1", "Lyz2", "Csf1r", "Cd14", "Fcgr3"),
  Macrophages = c("Adgre1", "Cd68", "C1qa", "C1qb", "C1qc"),
  M1s = c("Nos2"),
  M2s = c("Mrc1", "Arg1"),
  Cardiomyocytes = c("Tnnt2", "Myh6", "Tnnc1", "Tnni3"),
  Fibroblasts = c("DCN", "COL1A1", "COL1A2", "THY1"),
  Endothelials = c("TEK", "PECAM1", "FLT1", "VWF"),
  Mesothelials = c("MSLN", "UPK3B", "CALB2", "WT1"),
  SMCs = c("ACTA2", "TAGLN"),
  Epithelials = c("EPCAM", "KRT19", "CDH1", "KRT18"),
  Prolifs = c("MKI67"),
  name_use = "name_use",
  mycolor = c("#BC3C29FF", "#0072B5FF", "#E18727FF", "#20854EFF", "#7876B1FF",
    "#6F99ADFF", "#FFDC91FF", "#EE4C97FF", "#E64B35FF", "#4DBBD5FF", "#00A087FF",
    "#3C5488FF", "#F39B7FFF", "#8491B4FF", "#91D1C2FF", "#DC0000FF", "#7E6148FF",
    "#B09C85FF", "#3B4992FF", "#EE0000FF", "#008B45FF", "#631879FF", "#008280FF",
    "#BB0021FF", "#5F559BFF", "#A20056FF", "#808180FF", "#00468BFF", "#ED0000FF",
    "#42B540FF", "#0099B4FF", "#925E9FFF", "#FDAF91FF", "#AD002AFF", "#ADB6B6FF",
    "#374E55FF", "#DF8F44FF", "#00A1D5FF", "#B24745FF", 
     "#79AF97FF", "#6A6599FF",
    "#80796BFF", "#1f77b4", "#ff7f0e", "#279e68", "#d62728", "#aa40fc", "#8c564b",
    "#e377c2", "#b5bd61", "#17becf", "#aec7e8")
)
}
\arguments{
\item{scRNA}{Seurat object}

\item{label}{celltype or seurat_clusters or others}

\item{width}{pdf width}

\item{height}{pdfheight}

\item{mycolor}{mycolor}
}
\value{
pdf and ggplot object
}
\description{
bplot_scmarkers_mouse
}
