% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capitalize.R
\name{capitalize_first_letter}
\alias{capitalize_first_letter}
\title{capitalize_first_letter}
\usage{
capitalize_first_letter(str)
}
\arguments{
\item{str}{character string}
}
\value{
character string
}
\description{
capital the first letter, lowercase the rest of the letters
}
\examples{
# examples
str = c("aaa","bBB","CcC","dDd")
str = capitalize_first_letter(str)
print(str)
}
