% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert_to_capitalize_first_letter}
\alias{convert_to_capitalize_first_letter}
\title{convert_to_capitalize_first_letter}
\usage{
convert_to_capitalize_first_letter(str)
}
\arguments{
\item{str}{character string}
}
\value{
character string
}
\description{
capital the first letter, lowercase the rest of the letters
}
\examples{
# examples
str = c("aaa","bBB","CcC","dDd")
str = convert_to_capitalize_first_letter(str)
print(str)
}
