% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert_to_factor_by_column}
\alias{convert_to_factor_by_column}
\title{convert_to_factor_by_column}
\usage{
convert_to_factor_by_column(data_frame, columns = NULL)
}
\arguments{
\item{data_frame}{data frame}

\item{columns}{data frame}
}
\value{
numeric matrix
}
\description{
factor the data frame
}
\examples{
# examples
set.seed(123)
counts_matrix = matrix(rpois(500, lambda = 10), nrow = 3, ncol = 50)
colnames(counts_matrix) = paste0("sample", 1:ncol(counts_matrix))
rownames(counts_matrix) = c("PDCD1","CD274","MKI67")
result = convert_to_numeric_by_column(counts_matrix)
print(result)
}
