% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{cor_test_row}
\alias{cor_test_row}
\title{cor_test_row}
\usage{
cor_test_row(df, varible, methods = "pearson")
}
\arguments{
\item{df}{data frame}

\item{varible}{Gene Name}

\item{methods}{pearson or spearman}
}
\value{
data frame
}
\description{
Calculate the correlation between one gene and all genes (every row)
}
\examples{
# examples
set.seed(123)
n_rows = 100
n_cols = 100
df = data.frame(matrix(rnorm(n_rows * n_cols), nrow = n_rows))
rownames(df) = paste0("row", 1:n_rows)
colnames(df) = paste0("col", 1:n_cols)
result = cor_test_row(df, "col1", methods = "pearson")
print(result)
}
