% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts2tpm.R
\name{counts2tpm}
\alias{counts2tpm}
\title{counts2tpm}
\usage{
counts2tpm(counts)
}
\arguments{
\item{counts}{counts matrix, rownames is genes, colnames is samples}
}
\value{
TPM matrix
}
\description{
Standardization of the counts matrix to TPM
}
\examples{
# examples
set.seed(123)
counts_matrix = matrix(rpois(500, lambda = 10), nrow = 3, ncol = 50)
colnames(counts_matrix) = paste0("sample", 1:ncol(counts_matrix))
rownames(counts_matrix) = c("PDCD1","CD274","MKI67")
result = counts2tpm(counts_matrix)
print(result)
}
