% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cplot.R
\name{cplot2}
\alias{cplot2}
\title{cplot2}
\usage{
cplot2(data, GeneName1, GeneName2, DatasetName, method = "pearson")
}
\arguments{
\item{GeneName1}{GeneName1}

\item{GeneName2}{GeneName2}

\item{DatasetName}{DatasetName}

\item{method}{pearson or spearman}
}
\value{
pdf plot
}
\description{
expression correlation analysis
}
