% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma.R
\name{customized}
\alias{customized}
\title{customized}
\usage{
customized(
  geneset = c("IL2", "GZMA", "GNLY", "PRF1", "GZMB", "GZMK", "IFNG", "NKG7")
)
}
\arguments{
\item{geneset}{Gene Names/symbols}
}
\value{
customized results
}
\description{
customized
}
