% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoget.R
\name{downloadGSE}
\alias{downloadGSE}
\title{Download GEO Files}
\usage{
downloadGSE(gse_id, series = "matrix")
}
\arguments{
\item{gse_id}{Character. The GEO series ID.}

\item{series}{Character. The type of series to download ('matrix' or 'suppl').}
}
\value{
No return value. Files are downloaded to the current working directory.
}
\description{
Downloads GEO series files from NCBI's FTP server.
}
\examples{
downloadGSE('GSE129516', series = 'matrix')

}
