% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplot.R
\name{dplot1}
\alias{dplot1}
\title{dplot1}
\usage{
dplot1(
  data,
  Type = "Type",
  variable,
  levels = NULL,
  test.methods = "wilcox.test",
  DatasetName = "DatasetName",
  width = 4.5,
  height = 4,
  alphas = 0.5,
  mycolor = c("#BC3C29FF", "#0072B5FF", "#E18727FF", "#20854EFF", "#7876B1FF",
    "#6F99ADFF", "#FFDC91FF", "#EE4C97FF", "#E64B35FF", "#4DBBD5FF", "#00A087FF",
    "#3C5488FF", "#F39B7FFF", "#8491B4FF", "#91D1C2FF", "#DC0000FF", "#7E6148FF",
    "#B09C85FF", "#3B4992FF", "#EE0000FF", "#008B45FF", "#631879FF", "#008280FF",
    "#BB0021FF", "#5F559BFF", "#A20056FF", "#808180FF", "#00468BFF", "#ED0000FF",
    "#42B540FF", "#0099B4FF", "#925E9FFF", "#FDAF91FF", "#AD002AFF", "#ADB6B6FF",
    "#374E55FF", "#DF8F44FF", "#00A1D5FF", "#B24745FF", 
     "#79AF97FF", "#6A6599FF",
    "#80796BFF", "#1f77b4", "#ff7f0e", "#279e68", "#d62728", "#aa40fc", "#8c564b",
    "#e377c2", "#b5bd61", "#17becf", "#aec7e8")
)
}
\arguments{
\item{data}{data frame}

\item{Type}{data group}

\item{variable}{Variables or gene names analyzed}

\item{levels}{The order of arrangement in grouping visualization}

\item{test.methods}{The method of difference analysis "wilcox.test" or "kruskal.test"}

\item{DatasetName}{The Dataset Name of the data}

\item{width}{Width of pdf plot}

\item{height}{height of pdf plot}

\item{alphas}{Transparency of color matching}

\item{mycolor}{my color}
}
\value{
pdf plot
}
\description{
Differentially expression analysis
}
\examples{
# examples
set.seed(123)
n = 150
data = data.frame(Type = sample(c("A", "B", "C"), n, replace = TRUE),
                  variable1 = rnorm(n)
                  )
   dplot1(data,
          Type = "Type",
          variable = "variable1",
          test.methods = "kruskal.test",
          DatasetName = "MyDataset",
          levels = c("C","B","A"),
          width = 6,
          height = 5)
}
