% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drivergene.R
\name{drivergene}
\alias{drivergene}
\title{drivergene}
\usage{
drivergene(
  GeneName,
  binary = "C:\\\\Program Files\\\\Mozilla Firefox\\\\firefox.exe",
  neverAsk.saveToDisk = "application\\\\octet-stream",
  myport = sample(1:10000, 1, replace = TRUE)
)
}
\arguments{
\item{binary}{firefox.exe is where}

\item{neverAsk.saveToDisk}{just for fun}

\item{myport}{port}
}
\value{
html
}
\description{
Gene Function find
}
\examples{
if(.Platform$OS.type == "windows"){
username <- Sys.getenv("USERNAME")
}else{
username <- system("whoami", intern = TRUE)
username <- trimws(username)
}

file.copy(from = paste0("C:\\\\Users\\\\",username,"\\\\Desktop\\\\fromto_java\\\\binman.7z"),
to = paste0("C:\\\\Users\\\\",username,"\\\\AppData\\\\Local\\\\binman.7z"))

if(!require("devtools",quietly = TRUE)){
install.packages("devtools")
}

if(!require("BiocManager",quietly = TRUE)){
install.packages("BiocManager")
}

BiocManager::install("limma")
BiocManager::install("magick")
devtools::install_github("grswsci/fromto")
# examples
drivergene(GeneName = "CDC20",
          binary = "C:\\\\Program Files\\\\Mozilla Firefox\\\\firefox.exe",
          neverAsk.saveToDisk = "application\\\\octet-stream",
          myport = sample(1:10000, 1, replace = TRUE)
}
