% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drivergsea.R
\name{drivergsea}
\alias{drivergsea}
\title{drivergsea}
\usage{
drivergsea(
  GeneSet,
  binary = "C:\\\\Program Files\\\\Mozilla Firefox\\\\firefox.exe",
  neverAsk.saveToDisk = "application\\\\octet-stream",
  myport = sample(1:10000, 1, replace = TRUE)
)
}
\arguments{
\item{binary}{firefox.exe is where}

\item{neverAsk.saveToDisk}{just for fun}

\item{myport}{port}
}
\value{
html
}
\description{
gsea Function find
}
