% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driverpubmed.R
\name{driverpubmed}
\alias{driverpubmed}
\title{driverpubmed}
\usage{
driverpubmed(..., year_start = "2021", year_end = "2024")
}
\arguments{
\item{...}{search terms}

\item{year_start}{you know bro}

\item{year_end}{you know bro}
}
\value{
html and text
}
\description{
publications find
}
\examples{
# examples
driverpubmed('Melanoma', 'TCGA', 'Signature')
driverpubmed_example() # or
}
