% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driverpubmed.R
\name{driverpubmed2}
\alias{driverpubmed2}
\title{driverpubmed2}
\usage{
driverpubmed2(
  ...,
  year_start = "2021",
  year_end = "2024",
  nMax = 100,
  abstract = T
)
}
\arguments{
\item{...}{search terms}

\item{year_start}{you know bro}

\item{year_end}{you know bro}

\item{nMax}{you know bro}

\item{abstract}{you know bro}
}
\value{
html and text
}
\description{
publications find
}
