% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_fromto.R
\name{exp_fromto}
\alias{exp_fromto}
\title{exp_fromto}
\usage{
exp_fromto(df)
}
\arguments{
\item{df}{data frame}
}
\value{
data frame
}
\description{
computes the exponential function.
}
\examples{
# examples
set.seed(123)
n_rows = 5
n_cols = 3
df = data.frame(
 var1 = rnorm(n_rows),
 var2 = runif(n_rows),
 var3 = rpois(n_rows, lambda = 5)
)
result_df = exp_fromto(df)
print(result_df)
}
