% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.R
\name{find_index}
\alias{find_index}
\title{find_index}
\usage{
find_index(x, ys)
}
\arguments{
\item{x}{character}

\item{ys}{characters}
}
\value{
index
}
\description{
Is it indexs
}
\examples{
# examples
ys = c("This is a sample text where we have a note saying 'replaced by ID12345'.",
      "This is a sample text where we have a note saying 'replaced by ID12346'.",
      "This is a sample text where we have a note saying 'replaced nby ID12345'.",
      "This is a sample text where we have a note saying 'replaced nby ID12346'.")
x = "replaced by ID"
find_index(x = x,ys = ys)
}
