% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.R
\name{find_indexs}
\alias{find_indexs}
\title{find_indexs}
\usage{
find_indexs(xs, ys)
}
\arguments{
\item{xs}{characters}

\item{ys}{characters}
}
\value{
index
}
\description{
Is it indexs
}
\examples{
# examples
df = data.frame(
y = c("This is a sample text where we have a note saying 'replaced by ID12345'.",
      "This is a sample text where we have a note saying 'replaced by ID12346'.",
      "This is a sample text where we have a note saying 'replaced nby ID12345'.",
      "This is a sample text where we have a note saying 'replaced nby ID12346'.")
x = c("replaced by ID",
      "replaced nby ID",
      "replaced nby ID",
      "replaced by ID")
)
find_indexs(xs = df$x,ys = df$y)
}
