% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fplot.R
\name{fplot2}
\alias{fplot2}
\title{fplot2}
\usage{
fplot2(data, nameplot, height = 10)
}
\arguments{
\item{data}{data frame}

\item{nameplot}{plot name}

\item{height}{pdf height}
}
\value{
forest plot
}
\description{
forest plot
}
