% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fromto.R
\name{fromtoupdate}
\alias{fromtoupdate}
\title{fromtoupdate}
\usage{
fromtoupdate(gene_matrix, avereps_use = TRUE)
}
\arguments{
\item{gene_matrix}{gene_matrix}

\item{avereps_use}{= TRUE or use FALSE}
}
\value{
data frame
}
\description{
Update gene names
}
\examples{
# examples
set.seed(123)
Oldname_matrix = matrix(rpois(500, lambda = 10), nrow = 3, ncol = 50)
colnames(Oldname_matrix) = paste0("sample", 1:ncol(Oldname_matrix))
rownames(Oldname_matrix) = c("hPD-1","PDL1","MIB-1")
Newname_matrix = fromtoupdate(Oldname_matrix)
print(Newname_matrix)
}
