% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoann.R
\name{geoann}
\alias{geoann}
\title{geoann}
\usage{
geoann(GPL_ID, GEO_ID)
}
\arguments{
\item{GPL_ID}{GPL ID}

\item{GEO_ID}{GEO ID}
}
\value{
data frame
}
\description{
The probe name was annotated as gene name.
}
\examples{
# examples
geoget(GEO_IDs = "GSE103668")
library(tidyverse)
library(R.utils)
list.files() \%>\% grepl(".txt.gz$", ., fixed = FALSE) \%>\% which() \%>\% list.files()[.] \%>\% gunzip(., remove = FALSE, overwrite = TRUE)
data = geoann(GPL_ID = "GPL570",GEO_ID = "GSE103668")
print(data)
}
