% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocli.R
\name{geocli}
\alias{geocli}
\title{geocli}
\usage{
geocli(GEO_ID)
}
\arguments{
\item{GEO_ID}{GEO ID}
}
\value{
data frame
}
\description{
Organizing GEO Clinical Documents
}
\examples{
# examples
geoget(GEO_IDs = "GSE103668")
library(tidyverse)
library(R.utils)
list.files() \%>\% grepl(".txt.gz$", ., fixed = FALSE) \%>\% which() \%>\% list.files()[.] \%>\% gunzip(., remove = FALSE, overwrite = TRUE)
clinical = geocli(GEO_ID = "GSE103668")
print(clinical)
}
