% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoget.R
\name{geoget}
\alias{geoget}
\title{geoget}
\usage{
geoget(GEO_IDs)
}
\arguments{
\item{GEO_ID}{GEO ID}
}
\value{
txt.gz
}
\description{
Download the GEO dataset from a Mirror
}
\examples{
# examples
geoget(GEO_IDs = "GSE103668")
}
