% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoget.R
\name{geoget2}
\alias{geoget2}
\title{geoget2}
\usage{
geoget2(GEO_IDs)
}
\arguments{
\item{GEO_ID}{GEO ID}
}
\value{
txt.gz
}
\description{
Download the GEO dataset from a NCBI
}
\examples{
# examples
geoget(GEO_IDs = "GSE103668")
}
