% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoget.R
\name{gse2url}
\alias{gse2url}
\title{Generate GEO Series URL}
\usage{
gse2url(gse_id, series = "matrix")
}
\arguments{
\item{gse_id}{Character. The GEO series ID.}

\item{series}{Character. The type of series to access ('matrix' or 'suppl').}
}
\value{
A character string containing the constructed URL.
}
\description{
Constructs the URL for accessing GEO series files from NCBI's FTP server.
}
\examples{
gse2url('GSE129516', series = 'matrix')

}
\keyword{internal}
