% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hplot.R
\name{hplot1}
\alias{hplot1}
\title{hplot1}
\usage{
hplot1(
  hplot_input,
  bar_color_use = "steelblue",
  legend_heatmap_use,
  legend_topann_use,
  cellwidth_use = 1,
  cellheight_use = 0.5,
  fontsize_use = 10,
  round_use = 3,
  topann_color_use = c("#BC3C29FF", "#0072B5FF", "#E18727FF", "#20854EFF", "#7876B1FF",
    "#6F99ADFF", "#FFDC91FF", "#EE4C97FF", "#E64B35FF", "#4DBBD5FF", "#00A087FF",
    "#3C5488FF", "#F39B7FFF", "#8491B4FF", "#91D1C2FF", "#DC0000FF", "#7E6148FF",
    "#B09C85FF", "#3B4992FF", "#EE0000FF", "#008B45FF", "#631879FF", "#008280FF",
    "#BB0021FF", "#5F559BFF", "#A20056FF", "#808180FF", "#00468BFF", "#ED0000FF",
    "#42B540FF", "#0099B4FF", "#925E9FFF", "#FDAF91FF", "#AD002AFF", "#ADB6B6FF",
    "#374E55FF", "#DF8F44FF", "#00A1D5FF", "#B24745FF", 
     "#79AF97FF", "#6A6599FF",
    "#80796BFF", "#1f77b4", "#ff7f0e", "#279e68", "#d62728", "#aa40fc", "#8c564b",
    "#e377c2", "#b5bd61", "#17becf", "#aec7e8"),
  heatmap_name_use = "heatmap_name_use"
)
}
\arguments{
\item{hplot_input}{data frame, x is variables, y is groups}

\item{legend_heatmap_use}{legend_heatmap_use}

\item{legend_topann_use}{legend_topann_use}

\item{cellwidth_use}{cellwidth_use}

\item{cellheight_use}{cellheight_use}

\item{round_use}{round_use}

\item{topann_color_use}{topann_color_use}

\item{heatmap_name_use}{heatmap_name_use}

\item{bar_color_usedata}{default is steelblue}
}
\value{
pdf
}
\description{
heatmap plot
}
