% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSel.R
\name{ichat_deepseek}
\alias{ichat_deepseek}
\title{ichat_deepseek}
\usage{
ichat_deepseek(
  system_prompt_use = "你是一个R语言、Python与生物信息学专家",
  model_use = "deepseek-chat",
  ask_use,
  api_key = "sk-81e0a52d66d249d4bbe69b6d4afff6e0",
  base_url_use = "https://api.deepseek.com/v1/chat/completions",
  api_key_use = "sk-81e0a52d66d249d4bbe69b6d4afff6e0",
  split_use = FALSE,
  top_p_use = 0.9,
  frequency_penalty_use = 0.5,
  presence_penalty_use = 0.3,
  max_tokens_use = 8192,
  temperature_use = 0,
  timeout_use = 180
)
}
\arguments{
\item{system_prompt_use}{system_prompt_use}

\item{model_use}{model_use}

\item{ask_use}{ask_use}

\item{api_key}{api_key}

\item{base_url_use}{base_url_use}

\item{api_key_use}{api_key_use}

\item{split_use}{split_use = FALSE}

\item{top_p_use}{top_p_use}

\item{frequency_penalty_use}{frequency_penalty_use}

\item{presence_penalty_use}{presence_penalty_use}

\item{max_tokens_use}{max_tokens_use}

\item{temperature_use}{temperature_use}

\item{timeout_use}{timeout_use}
}
\value{
ai_reply
}
\description{
ichat_deepseek
}
