% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{ifelse_row_1}
\alias{ifelse_row_1}
\title{ifelse_row_1}
\usage{
ifelse_row_1(df, col_add, col_name, row_var, row_var_if, row_var_else)
}
\arguments{
\item{df}{data frame}

\item{col_add}{col add name}

\item{row_var}{row varible}

\item{row_var_if}{row var if}

\item{row_var_else}{row var else}
}
\value{
df data frame
}
\description{
ifelse_row
}
