% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{ifelse_row_2}
\alias{ifelse_row_2}
\title{ifelse_row_2}
\usage{
ifelse_row_2(
  df,
  col_add,
  col_name,
  row_var,
  row_var_if,
  row_var2,
  row_var2_if,
  row_var2_else
)
}
\arguments{
\item{df}{data frame}

\item{col_add}{col add name}

\item{row_var}{row varible}

\item{row_var_if}{row var if}

\item{row_var2}{row varible2}

\item{row_var2_if}{row var2 if}

\item{row_var2_else}{row var2 else}
}
\value{
df data frame
}
\description{
ifelse_row_2
}
