% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSel.R
\name{iinput}
\alias{iinput}
\title{iinput}
\usage{
iinput(driver, js_use, input_use)
}
\arguments{
\item{driver}{driver}

\item{js_use}{js_use}

\item{input_use}{input_use}
}
\value{
iinput
}
\description{
iinput
}
