% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSel.R
\name{iinput2}
\alias{iinput2}
\title{iinput2}
\usage{
iinput2(driver, js_use, input_use)
}
\arguments{
\item{driver}{driver}

\item{js_use}{js_use}

\item{input_use}{input_use}
}
\value{
iinput2
}
\description{
iinput2
}
