% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSel.R
\name{ipage_nodes_change}
\alias{ipage_nodes_change}
\title{ipage_nodes_change}
\usage{
ipage_nodes_change(
  driver,
  js_use,
  change_pre = "<br>|<p>|<code>",
  change_post = "\\n",
  page_use = NULL
)
}
\arguments{
\item{driver}{driver}

\item{js_use}{js_use}

\item{change_pre}{change_pre}

\item{change_post}{change_post}

\item{page_use}{page_use}
}
\value{
ipage_nodes_change
}
\description{
ipage_nodes_change
}
