% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSel.R
\name{ipage_texts}
\alias{ipage_texts}
\title{ipage_texts}
\usage{
ipage_texts(driver, js_use, page_use = NULL)
}
\arguments{
\item{driver}{driver}

\item{js_use}{js_use}

\item{page_use}{page_use}
}
\value{
ipage_texts
}
\description{
ipage_texts
}
