% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/not_in.R
\name{is_ins}
\alias{is_ins}
\title{is_ins}
\usage{
is_ins(df, xs, ys)
}
\arguments{
\item{df}{data.frame}

\item{xs}{characters}

\item{ys}{characters}
}
\value{
TRUE or FALSE
}
\description{
Is it ins
}
\examples{
# examples
df = data.frame(
y = c("This is a sample text where we have a note saying 'replaced by ID12345'.",
      "This is a sample text where we have a note saying 'replaced by ID12346'.",
      "This is a sample text where we have a note saying 'replaced nby ID12345'.",
      "This is a sample text where we have a note saying 'replaced nby ID12346'.")
x = c("replaced by ID",
      "replaced nby ID",
      "replaced nby ID",
      "replaced by ID")
)
is_ins(df,xs = df$x,ys = df$y)
}
