% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/not_in.R
\name{is_integer}
\alias{is_integer}
\title{is_integer}
\usage{
is_integer(x)
}
\arguments{
\item{x}{number}
}
\value{
TRUE or FALSE
}
\description{
Is it integer
}
\examples{
# examples
is_integer(5)
is_integer(5.5)
}
