% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSel.R
\name{isleep}
\alias{isleep}
\title{isleep}
\usage{
isleep(runif_use = 1, min_use = 4, max_use = 8)
}
\arguments{
\item{runif_use}{runif_use = 1}

\item{min_use}{min_use = 4}

\item{max_use}{max_use = 8}
}
\value{
isleep
}
\description{
isleep
}
