% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/journal.R
\name{journal_abstract}
\alias{journal_abstract}
\title{journal_abstract}
\usage{
journal_abstract(docx = "journal_abstract.docx")
}
\arguments{
\item{docx}{journal_abstract.docx}
}
\value{
txt html
}
\description{
journal abstract
}
