% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/journal.R
\name{journal_editor}
\alias{journal_editor}
\title{journal_editor}
\usage{
journal_editor(
  editor,
  Research_Field1,
  Research_Field2,
  Research_Field3,
  section,
  email,
  institution,
  Biography,
  chief = TRUE
)
}
\arguments{
\item{editor}{Editor-}

\item{Research_Field1}{Research_Field1}

\item{Research_Field2}{Research_Field2}

\item{Research_Field3}{Research_Field3}

\item{section}{Cell Death Research}

\item{email}{email}

\item{institution}{institution}

\item{Biography}{Biography}

\item{chief}{TRUE/FALSE}
}
\value{
txt html
}
\description{
journal_editor
}
