% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/journal.R
\name{journal_figure_legends_png}
\alias{journal_figure_legends_png}
\title{journal_figure_legends_png}
\usage{
journal_figure_legends_png(divname, docx = "journal_figure_legends_png.docx")
}
\arguments{
\item{docx}{journal_figure_legends_png.docx}
}
\value{
txt html
}
\description{
journal figure legends png
}
