% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/journal.R
\name{journal_reference_numbers_no}
\alias{journal_reference_numbers_no}
\title{journal_reference_numbers_no}
\usage{
journal_reference_numbers_no(docx = "journal_reference_numbers_no.docx")
}
\arguments{
\item{docx}{journal_reference_numbers_no.docx}
}
\value{
txt html
}
\description{
journal reference
}
