% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/journal.R
\name{journal_table_jpg}
\alias{journal_table_jpg}
\title{journal_table_jpg}
\usage{
journal_table_jpg(divname, docx = "journal_table_jpg.docx")
}
\arguments{
\item{docx}{journal_table_jpg.docx}
}
\value{
txt html
}
\description{
journal_table_jpg
}
