% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmpot.R
\name{kmplot1}
\alias{kmplot1}
\title{kmplot1}
\usage{
kmplot1(data, minprop = 0.3, GeneName, CancerType, Timeunit = "year")
}
\arguments{
\item{data}{data frame}

\item{minprop}{mini-proportion}

\item{GeneName}{GeneName}

\item{CancerType}{CancerType}

\item{Timeunit}{year month data}
}
\value{
km plot
}
\description{
km plot and log-rank test
}
