% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma.R
\name{limma_gene}
\alias{limma_gene}
\title{limma_gene}
\usage{
limma_gene(data, GeneName, pct = 0.3, DatasetName = "DatasetName")
}
\arguments{
\item{data}{row is gene, column name is sample.}

\item{GeneName}{Gene Name/symbol}

\item{pct}{proportion}

\item{DatasetName}{Custom output file name}
}
\value{
csv and limma results
}
\description{
limma gene
}
