% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma.R
\name{limma_gene_gsea}
\alias{limma_gene_gsea}
\title{limma_gene_gsea}
\usage{
limma_gene_gsea(
  data,
  GeneName,
  DatasetName = "DatasetName",
  pct = 0.3,
  gmt = NULL,
  gmtrds
)
}
\arguments{
\item{data}{row is gene, column name is sample.}

\item{GeneName}{Gene Name/symbol}

\item{DatasetName}{Custom output file name}

\item{pct}{proportion}

\item{gmt}{gmt file}

\item{gmtrds}{gmtrds file}
}
\value{
csv and limma results
}
\description{
limma_gene_gsea
}
