% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma.R
\name{limma_gene_gsea_customized}
\alias{limma_gene_gsea_customized}
\title{limma_gene_gsea_customized}
\usage{
limma_gene_gsea_customized(
  data,
  GeneName,
  customized,
  pct = 0.3,
  DatasetName = "DatasetName"
)
}
\arguments{
\item{data}{row is gene, column name is sample.}

\item{GeneName}{Gene Name/symbol}

\item{customized}{customized file}

\item{pct}{proportion}

\item{DatasetName}{Custom output file name}
}
\value{
csv and limma results
}
\description{
limma_gene_gsea_customized
}
