% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoget.R
\name{listGEO}
\alias{listGEO}
\title{List GEO Files}
\usage{
listGEO(gse_id, series = "matrix", full = T)
}
\arguments{
\item{gse_id}{Character. The GEO series ID.}

\item{series}{Character. The type of series to access ('matrix', 'suppl', or 'supplementary').}

\item{full}{Logic. Whether to return full URLs.}
}
\value{
A character vector of file names or full URLs.
}
\description{
Lists GEO series files from NCBI's FTP server.
}
\examples{
listGEO('GSE129516', series = 'matrix', full = TRUE)

}
