% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoget.R
\name{listURL}
\alias{listURL}
\title{List files from a URL}
\usage{
listURL(url)
}
\arguments{
\item{url}{Character. The URL to fetch files from.}
}
\value{
A character vector containing file names matching the pattern.
}
\description{
Extracts and lists files from a specified URL that match a certain pattern.
}
\examples{
listURL(gse2url('GSE129516', series = 'matrix'))

}
\keyword{internal}
