% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driverpubmed.R
\name{makeSteps}
\alias{makeSteps}
\title{Generate Step Ranges}
\usage{
makeSteps(len, step)
}
\arguments{
\item{len}{Numeric. The maximum value of the sequence.}

\item{step}{Numeric. The step size for generating the sequence.}
}
\value{
A list. Each element in the list represents a range defined by the `step` size.
        If `step` is 1, the list contains repeated numbers from 1 to `len`.
}
\description{
Generates a list of step ranges based on a specified length and step size.
}
\examples{
makeSteps(10, step = 2)
makeSteps(10, step = 1)

}
