% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.R
\name{merge_col}
\alias{merge_col}
\title{merge_col}
\usage{
merge_col(data1, data2)
}
\arguments{
\item{data1}{data frame}

\item{data2}{data frame}
}
\value{
data frame
}
\description{
Merge by colnames
}
\examples{
# examples
set.seed(123)
n_rows = 10
n_cols1 = 3
data1 = data.frame(matrix(rnorm(n_rows * n_cols1), nrow = n_rows))
rownames(data1) = paste0("sample", 1:n_rows)
colnames(data1) = paste0("var", 1:n_cols1)
data1 = t(data1)

set.seed(456)
n_cols2 = 2
data2 = data.frame(matrix(runif(n_rows * n_cols2), nrow = n_rows))
rownames(data2) = paste0("sample", 1:n_rows)
colnames(data2) = paste0("var", (n_cols1 + 1):(n_cols1 + n_cols2))
data2 = t(data2)

result = merge_col(data1, data2)
print(result)
}
