% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.R
\name{merge_col_add}
\alias{merge_col_add}
\title{merge_col_add}
\usage{
merge_col_add(
  data1,
  data2,
  data1_var,
  data2_var_same_data1,
  data2_var_add_data1
)
}
\arguments{
\item{data1}{data frame}

\item{data2}{data frame}

\item{data1_var}{data1_var}

\item{data2_var_same_data1}{data2_var_same_data1}

\item{data2_var_add_data1}{data2_var_add_data1}
}
\description{
Merge add line from another data frame
}
