% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.R
\name{merge_plot}
\alias{merge_plot}
\title{merge_plot}
\usage{
merge_plot(type = "pdf", ncol = 1)
}
\arguments{
\item{type}{plot type, you can choose "pdf","jpg","png" and "tiff"}

\item{ncol}{How many pictures are there in each column}
}
\value{
ggplot object
}
\description{
Merge by plot
}
\examples{
# examples
set.seed(123)
n = 150
data = data.frame(Type = sample(c("A", "B", "C"), n, replace = TRUE),
                  variable1 = rnorm(n))
   dplot1(data,
          Type = "Type",
          variable = "variable1",
          test.methods = "kruskal.test",
          DatasetName = "Name1",
          levels = c("C","B","A"),
          width = 6,
          height = 5)
   dplot1(data,
          Type = "Type",
          variable = "variable1",
          test.methods = "kruskal.test",
          DatasetName = "Name2",
          levels = c("C","B","A"),
          width = 6,
          height = 5)
   dplot1(data,
          Type = "Type",
          variable = "variable1",
          test.methods = "kruskal.test",
          DatasetName = "Name3",
          levels = c("C","B","A"),
          width = 6,
          height = 5)
plot_merge = merge_plot(type = "pdf", ncol = 3)
ggsave2("Figure_Merge.pdf",height = 1,width = 3)
}
