% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minmax.R
\name{minmax_row}
\alias{minmax_row}
\title{minmax_row}
\usage{
minmax_row(x)
}
\arguments{
\item{x}{matrix, rownames is genes, colnames is samples}
}
\value{
minmax matrix
}
\description{
Standardization of the matrix to minmax score (every row)
}
\examples{
# examples
set.seed(123)
counts_matrix = matrix(rpois(500, lambda = 10), nrow = 3, ncol = 50)
colnames(counts_matrix) = paste0("sample", 1:ncol(counts_matrix))
rownames(counts_matrix) = c("PDCD1","CD274","MKI67")
result = minmax_row(counts_matrix)
print(result)
print(max(as.numeric(result[,1])))
print(min(as.numeric(result[,1])))
}
