% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numericize.R
\name{numericize_by_column}
\alias{numericize_by_column}
\title{numericize_by_column}
\usage{
numericize_by_column(data_frame, column = NULL)
}
\arguments{
\item{x}{matrix, rownames is samples, colnames is genes}
}
\value{
numeric matrix
}
\description{
Standardization of the matrix to zscore (every col)
}
\examples{
# examples
set.seed(123)
counts_matrix = matrix(rpois(500, lambda = 10), nrow = 3, ncol = 50)
colnames(counts_matrix) = paste0("sample", 1:ncol(counts_matrix))
rownames(counts_matrix) = c("PDCD1","CD274","MKI67")
result = numericize_by_column(counts_matrix)
print(result)
}
