% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order.R
\name{order_fromto}
\alias{order_fromto}
\title{order_fromto}
\usage{
order_fromto(df, variable)
}
\arguments{
\item{x}{data frame, rownames is samples, colnames is variables}
}
\value{
data frame
}
\description{
The data box is in ascending order according to one of the columns
}
\examples{
# examples
set.seed(123)
counts_matrix = matrix(rpois(500, lambda = 10), nrow = 3, ncol = 50)
colnames(counts_matrix) = paste0("sample", 1:ncol(counts_matrix))
rownames(counts_matrix) = c("PDCD1","CD274","MKI67")
counts_matrix = t(counts_matrix)
result = order_fromto(df = counts_matrix,variable = "PDCD1")
print(result)
}
