% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pplot.R
\name{pplot}
\alias{pplot}
\title{pplot}
\usage{
pplot(
  df,
  variable = c("Quantile", "Response_NR"),
  levels = c("R", "NR"),
  legend_use = "Response",
  plot_name_use = "plotname_use"
)
}
\arguments{
\item{df}{data frame}

\item{variable}{default = c("Quantile","Response_NR")}

\item{levels}{default = c("R","NR")}

\item{legend_use}{default = "Response"}

\item{plot_name_use}{default = "plotname_use"}
}
\value{
pdf
}
\description{
pplot
}
