% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparkle.R
\name{protein_exp_cptac_impute}
\alias{protein_exp_cptac_impute}
\title{protein_exp_cptac_impute}
\usage{
protein_exp_cptac_impute(CancerType)
}
\arguments{
\item{CancerType}{CancerType}
}
\value{
data frame of the cptac protein expression of all genes with impute NA value
}
\description{
Read the cptac protein expression of all genes with impute NA value
}
