% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparkle.R
\name{rna_exp_geo_singledataset_tumor}
\alias{rna_exp_geo_singledataset_tumor}
\title{rna_exp_geo_singledataset_tumor}
\usage{
rna_exp_geo_singledataset_tumor(CancerType, DatasetName)
}
\arguments{
\item{CancerType}{CancerType}

\item{DatasetName}{DatasetName}
}
\value{
data frame of the single dataset tumor expression of all genes
}
\description{
Read the single dataset tumor expression of all genes
}
