% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparkle.R
\name{rna_exp_pancancer}
\alias{rna_exp_pancancer}
\title{rna_exp_pancancer}
\usage{
rna_exp_pancancer()
}
\arguments{
\item{not}{require}
}
\value{
the data frame of pan-cancer matrix
}
\description{
Read the pan-cancer matrix
}
