% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparkle.R
\name{rna_exp_sc_gene}
\alias{rna_exp_sc_gene}
\title{rna_exp_sc_gene}
\usage{
rna_exp_sc_gene(scRNA, GeneName)
}
\arguments{
\item{scRNA}{seurat object}

\item{GeneName}{One Gene or multi-Genes}
}
\value{
data frame of the sc rna expression of One Gene or multi-Genes
}
\description{
get sc rna data of selected genes
}
