% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparkle.R
\name{rna_exp_sc_tumor}
\alias{rna_exp_sc_tumor}
\title{rna_exp_sc_tumor}
\usage{
rna_exp_sc_tumor(CancerType)
}
\arguments{
\item{CancerType}{CancerType}
}
\value{
data frame of the sc rna tumor expression of all genes
}
\description{
Read the sc rna tumor expression of all genes
}
