% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparkle.R
\name{rna_exp_singlecancer}
\alias{rna_exp_singlecancer}
\title{rna_exp_singlecancer}
\usage{
rna_exp_singlecancer(CancerType)
}
\arguments{
\item{CancerType}{CancerType}
}
\value{
the data frame of the single cancer tumor expression of all gene
}
\description{
Read the single cancer tumor expression of all gene
}
