% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparkle.R
\name{rna_exp_singlecancer_multigenes}
\alias{rna_exp_singlecancer_multigenes}
\title{rna_exp_singlecancer_multigenes}
\usage{
rna_exp_singlecancer_multigenes(CancerType, multigenes)
}
\arguments{
\item{CancerType}{CancerType}

\item{multigenes}{multigenes}
}
\value{
the data frame of the single cancer tumor expression of multigenes
}
\description{
Read the single cancer tumor expression of multigenes
}
