% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparkle.R
\name{rna_exp_st}
\alias{rna_exp_st}
\title{rna_exp_st}
\usage{
rna_exp_st(CancerType)
}
\arguments{
\item{CancerType}{CancerType}
}
\value{
data frame of the st rna expression of all genes
}
\description{
Read the st rna expression of all genes
}
