% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparkle.R
\name{rna_exp_st_gene}
\alias{rna_exp_st_gene}
\title{rna_exp_st_gene}
\usage{
rna_exp_st_gene(stRNA, GeneName)
}
\arguments{
\item{GeneName}{One Gene or multi-Genes}

\item{scRNA}{seurat object}
}
\value{
data frame of the st rna expression of One Gene or multi-Genes
}
\description{
get st rna data of selected genes
}
