% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparkle.R
\name{rna_gene_exp_pancancer_zscore}
\alias{rna_gene_exp_pancancer_zscore}
\title{rna_gene_exp_pancancer_zscore}
\usage{
rna_gene_exp_pancancer_zscore(GeneName)
}
\arguments{
\item{GeneName}{GeneName}
}
\value{
the data frame of pan-cancer zscore expression of one gene
}
\description{
Read the pan-cancer zscore expression of one gene
}
