% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparkle.R
\name{rna_gene_exp_singlecancer_zscore}
\alias{rna_gene_exp_singlecancer_zscore}
\title{rna_gene_exp_singlecancer_zscore}
\usage{
rna_gene_exp_singlecancer_zscore(GeneName, CancerType)
}
\arguments{
\item{GeneName}{GeneName}

\item{CancerType}{CancerType}
}
\value{
the data frame of single cancer zscore expression of one gene
}
\description{
Read the single cancer zscore expression of one gene
}
