% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparkle.R
\name{rna_gene_survival_pancancer}
\alias{rna_gene_survival_pancancer}
\title{rna_gene_survival_pancancer}
\usage{
rna_gene_survival_pancancer(GeneName)
}
\arguments{
\item{GeneName}{GeneName}
}
\value{
the data frame of single cancer zscore expression of one gene
}
\description{
Read the pan-cancer expression and survival information of one gene
}
