% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparkle.R
\name{rna_gene_survival_singlecancer}
\alias{rna_gene_survival_singlecancer}
\title{rna_gene_survival_singlecancer}
\usage{
rna_gene_survival_singlecancer(GeneName, CancerType)
}
\arguments{
\item{GeneName}{GeneName}

\item{CancerType}{CancerType}
}
\value{
data frame of the single cancer tumor expression and survival information of one gene
}
\description{
Read the single cancer tumor expression and survival information of one gene
}
