% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparkle.R
\name{rna_sample_pancancer}
\alias{rna_sample_pancancer}
\title{rna_sample_pancancer}
\usage{
rna_sample_pancancer()
}
\arguments{
\item{not}{require}
}
\value{
the data frame of pan-cancer sample
}
\description{
Read the pan-cancer sample
}
