% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparkle.R
\name{rna_survival_pancancer}
\alias{rna_survival_pancancer}
\title{rna_survival_pancancer}
\usage{
rna_survival_pancancer()
}
\arguments{
\item{not}{require}
}
\value{
the data frame of pan-cancer survival
}
\description{
Read the pan-cancer survival
}
