% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rplot.R
\name{rplot}
\alias{rplot}
\title{rplot}
\usage{
rplot(
  df,
  group = "Response_NR",
  varible_analysis,
  varible_control = "NR",
  main_name_use = "Immunotherapy response prediction",
  plot_name_use = "plotname_use"
)
}
\arguments{
\item{df}{data frame}

\item{group}{default = "Response_NR"}

\item{varible_analysis}{analysis which varible}

\item{varible_control}{in varible_analysis, control group name is what}

\item{main_name_use}{main name use}

\item{plot_name_use}{default = "plotname_use"}
}
\value{
pdf
}
\description{
rplot
}
