% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc.R
\name{sc_10x}
\alias{sc_10x}
\title{sc_10x}
\usage{
sc_10x(path, dir_name_idx = 1, sc_10x_name_idx = 4, str = "_")
}
\arguments{
\item{path}{working directory}

\item{dir_name_idx}{Sample name index}

\item{sc_10x_name_idx}{10x three files index}

\item{str}{What to divide it by}
}
\value{
10x three files
}
\description{
One key to prepare multiple samples of 10x three files
}
