% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc.R
\name{sc_name_cell_with_gene}
\alias{sc_name_cell_with_gene}
\title{sc_name_cell_with_gene}
\usage{
sc_name_cell_with_gene(
  scRNA,
  slot = "counts",
  gene_use,
  cell_use,
  label_use = "celltype",
  label_new = "CellType_New",
  method_use = "pn"
)
}
\arguments{
\item{scRNA}{Seurat Object with main label}

\item{slot}{default = "counts", or use "data"}

\item{gene_use}{gene symbol}

\item{cell_use}{cell name}

\item{label_use}{label in colnames(scRNA@meta.data)}

\item{label_new}{label_new in colnames(scRNA@meta.data)}

\item{method_use}{default =  "pn", means positive and negative. or use "med", means median}
}
\description{
name cell with gene
}
