% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc.R
\name{sc_qc_human}
\alias{sc_qc_human}
\title{sc_qc_human}
\usage{
sc_qc_human(
  scRNA,
  min_Gene = 500,
  max_Gene = 4000,
  max_UMI = 15000,
  pct_MT = 10,
  pct_RB = 1,
  plot_featrures = c("nFeature_RNA", "nCount_RNA", "percent.MT", "percent.RB"),
  group = "orig.ident",
  width = 28,
  height = 8,
  mycolor = c("#BC3C29FF", "#0072B5FF", "#E18727FF", "#20854EFF", "#7876B1FF",
    "#6F99ADFF", "#FFDC91FF", "#EE4C97FF", "#E64B35FF", "#4DBBD5FF", "#00A087FF",
    "#3C5488FF", "#F39B7FFF", "#8491B4FF", "#91D1C2FF", "#DC0000FF", "#7E6148FF",
    "#B09C85FF", "#3B4992FF", "#EE0000FF", "#008B45FF", "#631879FF", "#008280FF",
    "#BB0021FF", "#5F559BFF", "#A20056FF", "#808180FF", "#00468BFF", "#ED0000FF",
    "#42B540FF", "#0099B4FF", "#925E9FFF", "#FDAF91FF", "#AD002AFF", "#ADB6B6FF",
    "#374E55FF", "#DF8F44FF", "#00A1D5FF", "#B24745FF", 
     "#79AF97FF", "#6A6599FF",
    "#80796BFF", "#1f77b4", "#ff7f0e", "#279e68", "#d62728", "#aa40fc", "#8c564b",
    "#e377c2", "#b5bd61", "#17becf", "#aec7e8")
)
}
\arguments{
\item{scRNA}{working directory}

\item{min_Gene}{min_Gene default = 500}

\item{max_Gene}{max_Gene default = 4000}

\item{max_UMI}{max_UMI default = 15000}

\item{plot_featrures}{plot_featrures default = c("nFeature_RNA","nCount_RNA","percent.MT","percent.RB")}

\item{group}{group default = "orig.ident"}

\item{width}{width = 28}

\item{height}{height = 8}

\item{mycolor}{mycolor}

\item{pct_mt}{pct_mt default = 10}

\item{pct_rb}{pct_rb default = 1}
}
\value{
Seurat Object
}
\description{
Seurat Object
}
