% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scpathway.R
\name{sc_run_AddModuleScore}
\alias{sc_run_AddModuleScore}
\title{sc_run_AddModuleScore}
\usage{
sc_run_AddModuleScore(scRNA, markers, name_use = "AddModuleScore")
}
\arguments{
\item{scRNA}{Seurat Object}

\item{markers}{markers}

\item{name_use}{default = 'AddModuleScore'}
}
\description{
sc run AddModuleScore
}
