% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scpathway.R
\name{sc_run_AUCell}
\alias{sc_run_AUCell}
\title{sc_run_AUCell}
\usage{
sc_run_AUCell(scRNA, markers, cores_use = 2)
}
\arguments{
\item{scRNA}{Seurat Object}

\item{markers}{markers}

\item{cores_use}{default = 2}
}
\description{
sc run AUCell
}
