% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellchat.R
\name{sc_run_cellchat}
\alias{sc_run_cellchat}
\title{sc_run_cellchat}
\usage{
sc_run_cellchat(
  scRNA,
  gene_use = NULL,
  cell_use,
  label_use = "CellType_New",
  is_v5 = FALSE,
  height_1 = 8,
  height_2 = 8
)
}
\arguments{
\item{scRNA}{Seurat Object}

\item{gene_use}{gene symbol, default = NULL}

\item{cell_use}{cell name}

\item{label_use}{label in colnames(scRNA@meta.data)}

\item{is_v5}{is v5 Seurat Object default = FALSE}

\item{height_1}{default = 8}

\item{height_2}{default = 8}
}
\value{
cellchat Object
}
\description{
Seurat Object
}
