% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc.R
\name{sc_run_clustree}
\alias{sc_run_clustree}
\title{sc_run_clustree}
\usage{
sc_run_clustree(
  scRNA,
  from = 0.1,
  to = 1,
  seqs = 0.1,
  reduction_use = "harmony",
  height = 15,
  width = 10
)
}
\arguments{
\item{scRNA}{Seurat Object}

\item{from}{default = 0.1}

\item{to}{default = 1}

\item{seqs}{default = 0.1}

\item{reduction_use}{default = 0.1}

\item{height}{default = 15}

\item{width}{default = 10}
}
\description{
sc_run_clustree
}
