% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc.R
\name{sc_run_DoubletFinder}
\alias{sc_run_DoubletFinder}
\title{sc_run_DoubletFinder}
\usage{
sc_run_DoubletFinder(scRNA, if_sct = FALSE)
}
\arguments{
\item{scRNA}{Seurat Object}

\item{if_sct}{default = FALSE}
}
\description{
Seurat Object
}
