% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scpathway.R
\name{sc_run_UCell}
\alias{sc_run_UCell}
\title{sc_run_UCell}
\usage{
sc_run_UCell(scRNA, markers, cores_use = 2, name_use = "")
}
\arguments{
\item{scRNA}{Seurat Object}

\item{markers}{markers}

\item{cores_use}{default = 2}

\item{name_use}{default = ""}
}
\description{
sc run UCell
}
