% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc.R
\name{sc_run_umap_tsne}
\alias{sc_run_umap_tsne}
\title{sc_run_umap_tsne}
\usage{
sc_run_umap_tsne(scRNA, reduction_use = "harmony", resolution = 0.8)
}
\arguments{
\item{scRNA}{Seurat Object}

\item{reduction_use}{default = "pca"}

\item{resolution}{default = 0.8}
}
\description{
Seurat Object
}
