% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc.R
\name{sc_scale}
\alias{sc_scale}
\title{sc_scale}
\usage{
sc_scale(scRNA, methods_my = NA, if_vars_to_regress = "TRUE")
}
\arguments{
\item{scRNA}{Seurat Object}

\item{methods_my}{methods_my default = NA, or set as "sct"}

\item{if_vars_to_regress}{if_vars_to_regress default = "TRUE", or set as "FALSE"}
}
\description{
Seurat Object
}
