% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc.R
\name{sc_tisch2_reanalysis}
\alias{sc_tisch2_reanalysis}
\title{sc_tisch2_reanalysis}
\usage{
sc_tisch2_reanalysis(
  scRNA,
  nfeatures = 100,
  cell_use,
  label_use = "CellType_FineLabel",
  resolution_use = 1,
  reduction_use = "pca"
)
}
\arguments{
\item{scRNA}{Seurat Object with main label}

\item{nfeatures}{default = 100}

\item{cell_use}{cell name}

\item{label_use}{label in colnames(scRNA@meta.data)}

\item{resolution_use}{default = 1}

\item{reduction_use}{default = "pca"}
}
\description{
name cell with gene
}
