% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmarkers.R
\name{scmarkers}
\alias{scmarkers}
\title{scmarkers}
\usage{
scmarkers(label = "main_lable")
}
\arguments{
\item{label}{"main_lable" "fine_lable_t" "fine_lable_b" "fine_lable_dc" "fine_lable_myeloid" "fine_lable_fibroblast" "fine_lable_endothelial" "fine_lable_epithelial"}
}
\value{
scmarkers
}
\description{
scmarkers in publications
}
\examples{
# examples
scmarkers(label = "main_lable")
scmarkers(label = "fine_lable_t")
scmarkers(label = "fine_lable_b")
scmarkers(label = "fine_lable_dc")
scmarkers(label = "fine_lable_myeloid")
scmarkers(label = "fine_lable_fibroblast")
scmarkers(label = "fine_lable_endothelial")
scmarkers(label = "fine_lable_epithelial")
}
