% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st.R
\name{st_read_10x_3file_hires}
\alias{st_read_10x_3file_hires}
\title{st_read_10x_3file_hires}
\usage{
st_read_10x_3file_hires(data_dir_1, data_dir_2, CancerName)
}
\arguments{
\item{data_dir_1}{data_dir_1}

\item{data_dir_2}{data_dir_2}

\item{CancerName}{CancerName}
}
\value{
seurat object
}
\description{
st_read_10x_3file_hires
}
