% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st.R
\name{st_read_10x_h5_hires}
\alias{st_read_10x_h5_hires}
\title{st_read_10x_h5_hires}
\usage{
st_read_10x_h5_hires(CancerName, path = "/DL")
}
\arguments{
\item{CancerName}{CancerName}

\item{path}{path}
}
\value{
seurat object
}
\description{
st_read_10x_h5_hires
}
