% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st.R
\name{st_read_10x_h5_lowres}
\alias{st_read_10x_h5_lowres}
\title{st_read_10x_h5_lowres}
\usage{
st_read_10x_h5_lowres(CancerName, path = "/DL")
}
\arguments{
\item{CancerName}{CancerName}

\item{path}{path}
}
\value{
seurat object
}
\description{
st_read_10x_h5_lowres
}
