% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{subset_row1}
\alias{subset_row1}
\title{subset_row1}
\usage{
subset_row1(df, col_name, row_var)
}
\arguments{
\item{df}{data frame}

\item{col_name}{col name}

\item{row_var}{row varible}
}
\value{
df data frame
}
\description{
subset row1
}
