% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparkle.R
\name{survival_filter_month}
\alias{survival_filter_month}
\title{survival_filter_month}
\usage{
survival_filter_month(data, SurvivalType)
}
\arguments{
\item{data}{data frame}

\item{SurvivalType}{SurvivalType OS DSS DFI PFI}
}
\value{
zscore result
}
\description{
survival filter month
}
