% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cox.R
\name{survival_multicox}
\alias{survival_multicox}
\title{survival_multicox}
\usage{
survival_multicox(data)
}
\arguments{
\item{data}{data frame}
}
\value{
multicox results
}
\description{
The first column is the survival time, the second is the survival state, and the third and all subsequent columns are the variables to be analyzed.
}
