% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uplot.R
\name{uplot_gene2}
\alias{uplot_gene2}
\title{uplot_gene2}
\usage{
uplot_gene2(scRNA, marker, dim = "umap", size = 0.8, ncol = NULL)
}
\arguments{
\item{scRNA}{Seurat object}

\item{marker}{one gene or many genes}

\item{dim}{umap or tsne visualization}

\item{size}{dot size}

\item{ncol}{Multiple genes are arranged in columns and how many columns u want}
}
\value{
ggplot object
}
\description{
umap or tsne
}
