% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uplot.R
\name{uplot_gene3}
\alias{uplot_gene3}
\title{uplot_gene3}
\usage{
uplot_gene3(
  scRNA,
  marker,
  slot_use = "counts",
  reduction_use = "umap",
  size = 0.8,
  nameplot = "name"
)
}
\arguments{
\item{scRNA}{Seurat object}

\item{marker}{one gene}

\item{reduction_use}{umap or tsne}

\item{slot}{counts or data}

\item{ncol}{Multiple genes are arranged in columns and how many columns u want}
}
\value{
ggplot object
devtools::install_github("powellgenomicslab/Nebulosa")
}
\description{
umap or tsne
}
