% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/with.R
\name{withBackground}
\alias{withBackground}
\title{withBackground}
\usage{
withBackground(expr)
}
\arguments{
\item{expr}{raw expression}

\item{env}{caller environment}
}
\description{
Execute the expression in the background.
}
