% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/with.R
\name{withIn}
\alias{withIn}
\title{with function cluster}
\usage{
withIn(expr, env = parent.frame(), ...)
}
\arguments{
\item{expr}{raw expression}

\item{env}{caller environment}

\item{...}{Temporary variable}
}
\value{
what expr returns
}
\description{
Similar to function decorators, but more flexible, the advantages are:
1. No need to know the parameters of the original function;
2. Allow free transfer of the raw expression
Actually, it's like the 'with' statement in Python.
}
