% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/with.R
\name{withMessage}
\alias{withMessage}
\title{withMessage}
\usage{
withMessage(expr, text, coloredCat = cat, env = parent.frame(), ...)
}
\arguments{
\item{expr}{raw expression}

\item{text}{message}

\item{env}{caller environment}

\item{...}{Temporary variable}
}
\description{
Prompt message after the expression is executed.
}
