% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/with.R
\name{withNothing}
\alias{withNothing}
\title{withNothing}
\usage{
withNothing(expr, env = parent.frame(), ...)
}
\arguments{
\item{expr}{raw expression}

\item{env}{caller environment}

\item{...}{Temporary variable}
}
\value{
what expr returns
}
\description{
Do nothing, only used for testing the withIn function.
}
\examples{
a = 3
b = 4
fun = function() {
a = 1
b = 2
withNothing(a+b)
}
print(fun()) # 3
print(withNothing(a + b)) # 7
print(withNothing(a + b, a = 4, b = 5)) # 9

}
