% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/with.R
\name{withNull}
\alias{withNull}
\title{with}
\usage{
withNull(expr, env = environment())
}
\arguments{
\item{expr}{raw expression}

\item{env}{caller environment}
}
\value{
what expr returns
}
\description{
Similar to function decorators, but more flexible, the advantages are:
1. No need to know the parameters of the original function;
2. Allow free transfer of the raw expression
Actually, it's like the 'with' statement in Python.
}
\examples{
a = 3
b = 4
fun = function() {
a = 1
b = 2
withNull(a+b)
}
fun() # 3
withNull(a+b) # 7

}
