% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/with.R
\name{withSessions}
\alias{withSessions}
\title{withSessions}
\usage{
withSessions(expr, nWorker = 4, env = environment())
}
\arguments{
\item{expr}{raw expression}

\item{nWorker}{n of sessions}

\item{env}{caller environment}
}
\description{
First open multiple threads, then close.
}
