% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/with.R
\name{withSleep}
\alias{withSleep}
\title{withSleep}
\usage{
withSleep(expr, maxSec = 10, env = environment())
}
\arguments{
\item{expr}{raw expression}

\item{maxSec}{max sleep time(in seconds)}

\item{env}{caller environment}
}
\description{
First change the working directory to a new path, then change it back.
}
