% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zscore.R
\name{zscore_col}
\alias{zscore_col}
\title{zscore_col}
\usage{
zscore_col(x)
}
\arguments{
\item{x}{matrix, rownames is samples, colnames is genes}
}
\value{
zscore matrix
}
\description{
Standardization of the matrix to zscore (every col)
}
\examples{
# examples
set.seed(123)
counts_matrix = matrix(rpois(500, lambda = 10), nrow = 3, ncol = 50)
colnames(counts_matrix) = paste0("sample", 1:ncol(counts_matrix))
rownames(counts_matrix) = c("PDCD1","CD274","MKI67")
counts_matrix = t(counts_matrix)
result = zscore_col(counts_matrix)
print(result)
print(var(as.numeric(result[,1])))
print(mean(as.numeric(result[,1])))
}
