% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparkle.R
\name{zscore_col_genename}
\alias{zscore_col_genename}
\title{zscore_col_genename}
\usage{
zscore_col_genename(data, GeneName, FilterOut = TRUE)
}
\arguments{
\item{data}{data frame}

\item{GeneName}{GeneName}

\item{FilterOut}{TRUE is to remove outliers, or use False}
}
\value{
zscore result
}
\description{
zscore Designated Genes
}
