% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zscore.R
\name{zscore_row}
\alias{zscore_row}
\title{zscore_row}
\usage{
zscore_row(x)
}
\arguments{
\item{x}{matrix, rownames is genes, colnames is samples}
}
\value{
zscore matrix
}
\description{
Standardization of the matrix to zscore (every row)
}
\examples{
# examples
set.seed(123)
counts_matrix = matrix(rpois(500, lambda = 10), nrow = 3, ncol = 50)
colnames(counts_matrix) = paste0("sample", 1:ncol(counts_matrix))
rownames(counts_matrix) = c("PDCD1","CD274","MKI67")
result = zscore_row(counts_matrix)
print(result)
print(var(as.numeric(result[1,])))
print(mean(as.numeric(result[1,])))
}
