% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/not_in.R
\name{is_in}
\alias{is_in}
\title{is_integer}
\usage{
is_in(x, y)
}
\arguments{
\item{x}{number}
}
\value{
TRUE or FALSE
}
\description{
Is it integer
}
\examples{
# examples
y = "This is a sample text where we have a note saying 'replaced by ID12345'."
x = "replaced by ID"
is_in(x,y)
}
