% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/journal.R
\name{journal_editor}
\alias{journal_editor}
\title{journal_editor}
\usage{
journal_editor(editor, section, email, institution, focus, chief = TRUE)
}
\arguments{
\item{editor}{Editor-}

\item{section}{Cell Death Research}

\item{email}{email}

\item{institution}{institution}

\item{focus}{focus}

\item{chief}{TRUE/FALSE}
}
\value{
txt html
}
\description{
journal_editor
}
