% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc.R
\name{sc_10x_seurat}
\alias{sc_10x_seurat}
\title{sc_10x_seurat}
\usage{
sc_10x_seurat(path, if_join = FALSE)
}
\arguments{
\item{path}{working directory}

\item{if_join}{if Join Layers}
}
\value{
Seurat Object
}
\description{
One key to Create Seurat Object
}
