% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/with.R
\name{withMessage}
\alias{withMessage}
\title{withMessage}
\usage{
withMessage(expr, ..., sep = " ", coloredCat = cat, env = environment())
}
\arguments{
\item{expr}{raw expression}

\item{env}{caller environment}

\item{text}{message}
}
\description{
Prompt message after the expression is executed.
}
