% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/with.R
\name{withPath}
\alias{withPath}
\title{withPath}
\usage{
withPath(expr, path, env = environment())
}
\arguments{
\item{expr}{raw expression}

\item{path}{new working directory}

\item{env}{caller environment}
}
\description{
First change the working directory to a new path, then change it back.
}
\examples{
dir.create('tmp')
withPath(write.csv(iris, 'iris.csv'), 'tmp')
list.files('tmp') # iris.csv
}
